+++ src/actions/index.js
@@ -5,6 +5,6 @@ export const REQUEST_NEWS = 'REQUEST_NEWS';
 export const RECEIVE_NEWS = 'RECEIVE_NEWS';
 export const MARK_NOTIFICATION_SEEN = 'MARK_NOTIFICATION_SEEN';
 
-export const FETCH_ERROR = 'FETCH_ERROR'; 
+export const FETCH_ERROR = 'FETCH_ERROR';
 
 export const CLEAR_ITEM_ERRORS = 'CLEAR_ITEM_ERRORS';
